+++ Task.Connector/Connectors/PostgreConnector.cs
@@ -1,5 +1,6 @@
 ﻿using Dapper;
 using Npgsql;
+using Task.Connector.Exceptions;
 using Task.Connector.Interfaces;
 using Task.Connector.Models;
 using Task.Integration.Data.Models;
@@ -17,6 +18,9 @@ namespace Task.Connector.Connectors
 
         public void AddUserPermissions(string userLogin, IEnumerable<string> rightIds)
         {
+            if (IsUserExists(userLogin))
+                throw new UserNotFoundException(userLogin);
+
             var rightQuery = "INSERT INTO \"TestTaskSchema\".\"UserRequestRight\" (\"userId\", \"rightId\")" +
                 "VALUES (@Login, @RightId)";
 
@@ -99,6 +103,9 @@ namespace Task.Connector.Connectors
 
         public IEnumerable<string> GetUserPermissions(string userLogin)
         {
+            if (IsUserExists(userLogin))
+                throw new UserNotFoundException(userLogin);
+
             var query = "SELECT rr.\"name\" FROM \"TestTaskSchema\".\"UserRequestRight\" urr " +
                 "INNER JOIN \"TestTaskSchema\".\"RequestRight\" rr ON rr.\"id\" = urr.\"rightId\" " +
                 "WHERE urr.\"userId\" = 'GlavnyyNN' " +
@@ -112,6 +119,9 @@ namespace Task.Connector.Connectors
 
         public IEnumerable<UserProperty> GetUserProperties(string userLogin)
         {
+            if (IsUserExists(userLogin))
+                throw new UserNotFoundException(userLogin);
+
             var user = GetUserObjectProperty(userLogin);
 
             return user.GetProperties();
@@ -143,6 +153,9 @@ namespace Task.Connector.Connectors
 
         public void RemoveUserPermissions(string userLogin, IEnumerable<string> rightIds)
         {
+            if (IsUserExists(userLogin))
+                throw new UserNotFoundException(userLogin);
+
             var rightQuery = "DELETE FROM \"TestTaskSchema\".\"UserRequestRight\" " +
                 "WHERE \"userId\" = @Login and \"rightId\" = ANY(@RightIds);";
 
@@ -167,6 +180,9 @@ namespace Task.Connector.Connectors
 
         public void UpdateUserProperties(IEnumerable<UserProperty> properties, string userLogin)
         {
+            if (IsUserExists(userLogin))
+                throw new UserNotFoundException(userLogin);
+
             var user = GetUserObjectProperty(userLogin);
 
             user.UpdateProperties(properties);
